stripe_test/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.10
idna==3.6
packaging==23.2
python-dotenv==1.0.1
requests==2.31.0
sqlparse==0.4.4
stripe==8.4.0
typing_extensions==4.10.0
urllib3==2.2.1
